Rename script from  warpMemoryFinalizer to returnMemoryFinalizer
