+++ src/passes/index.ts
@@ -35,4 +35,4 @@ export * from './unreachableStatementPruner';
 export * from './usingForResolver';
 export * from './variableDeclarationExpressionSplitter';
 export * from './variableDeclarationInitialiser';
-export * from './warpMemoryFinalizer';
+export * from './returnMemoryFinalizer';
